change weigher return type to int. Rename statistic to getTotalWeight

+++ cache2k-api/src/main/java/org/cache2k/Weigher.java
@@ -34,6 +34,6 @@ public interface Weigher<K, V> {
    * @return a positive long value representing the relative weight in comparison to the other
    * entries in the cache.
    */
-  long weigh(K key, V value);
+  int weigh(K key, V value);
 
 }
